refactor: Pot 생성 및 지원 Service CQRS 구조로 리팩토링
