+++ src/main/java/stackpot/stackpot/pot/controller/PotApplicationController.java
@@ -7,9 +7,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import stackpot.stackpot.apiPayload.ApiResponse;
-import stackpot.stackpot.pot.service.PotApplicationService;
 import stackpot.stackpot.pot.dto.PotApplicationRequestDto;
 import stackpot.stackpot.pot.dto.PotApplicationResponseDto;
+import stackpot.stackpot.pot.service.PotApplicationCommandService;
+import stackpot.stackpot.pot.service.PotApplicationQueryService;
 
 import java.util.List;
 
@@ -19,7 +20,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PotApplicationController {
 
-    private final PotApplicationService potApplicationService;
+    private final PotApplicationCommandService potApplicationCommandService;
+    private final PotApplicationQueryService potApplicationQueryService;
     @Operation(summary = "팟 지원 API")
     @PostMapping
     public ResponseEntity<ApiResponse<PotApplicationResponseDto>> applyToPot(
@@ -27,14 +29,14 @@ public class PotApplicationController {
             @RequestBody @Valid PotApplicationRequestDto requestDto) {
 
         // 팟 지원 로직 호출
-        PotApplicationResponseDto responseDto = potApplicationService.applyToPot(requestDto, potId);
+        PotApplicationResponseDto responseDto = potApplicationCommandService.applyToPot(requestDto, potId);
 
         return ResponseEntity.ok(ApiResponse.onSuccess(responseDto)); // 성공 시 ApiResponse로 감싸서 응답 반환
     }
     @Operation(summary = "팟 지원 취소 API")
     @DeleteMapping
     public ResponseEntity<ApiResponse<Void>> cancelApplication(@PathVariable("pot_id") Long potId) {
-        potApplicationService.cancelApplication(potId);
+        potApplicationCommandService.cancelApplication(potId);
         return ResponseEntity.ok(ApiResponse.onSuccess(null));
     }
 
@@ -44,7 +46,7 @@ public class PotApplicationController {
     public ResponseEntity<ApiResponse<List<PotApplicationResponseDto>>> getApplicants(
             @PathVariable("pot_id") Long potId) {
         // 서비스 호출
-        List<PotApplicationResponseDto> applicants = potApplicationService.getApplicantsByPotId(potId);
+        List<PotApplicationResponseDto> applicants = potApplicationQueryService.getApplicantsByPotId(potId);
 
         return ResponseEntity.ok(ApiResponse.onSuccess(applicants)); // 성공 시 ApiResponse로 감싸서 응답 반환
     }
