+++ src/main/java/stackpot/stackpot/pot/service/PotApplicationQueryService.java
@@ -1,0 +1,13 @@
+package stackpot.stackpot.pot.service;
+
+import stackpot.stackpot.pot.dto.PotApplicationResponseDto;
+import stackpot.stackpot.pot.dto.PotDetailWithApplicantsResponseDto;
+
+import java.util.List;
+
+public interface PotApplicationQueryService {
+
+    List<PotApplicationResponseDto> getApplicantsByPotId(Long potId);
+
+    PotDetailWithApplicantsResponseDto getPotDetailsAndApplicants(Long potId);
+}
